Add logging verbosity control
Log all comments received. With --verbose enabled, you can see these on
the console output.

Add a timeout so that the application will abort if no data is seen on
the bus for 60 seconds. If set to auto-restart (e.g. by systemd), DTR
will be pulled low on exit, which resets the Arduino, so this option
acts like a watchdog for the system.